@@ -31,66 +31,66 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@babel/core": "^7.4.3",
-    "@babel/plugin-proposal-class-properties": "^7.4.0",
+    "@babel/core": "^7.7.2",
+    "@babel/plugin-proposal-class-properties": "^7.7.0",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-transform-runtime": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/runtime-corejs3": "^7.4.5",
-    "@svgr/webpack": "^4.3.1",
+    "@babel/plugin-transform-runtime": "^7.6.2",
+    "@babel/preset-env": "^7.7.1",
+    "@babel/preset-react": "^7.7.0",
+    "@babel/runtime-corejs3": "^7.7.2",
+    "@svgr/webpack": "^4.3.3",
     "add-asset-html-webpack-plugin": "^3.1.3",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.5",
-    "babel-plugin-import": "^1.12.0",
+    "babel-eslint": "^10.0.3",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-import": "^1.12.2",
     "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^5.0.2",
-    "cross-env": "^5.2.0",
-    "css-loader": "^3.0.0",
+    "copy-webpack-plugin": "^5.0.5",
+    "cross-env": "^6.0.3",
+    "css-loader": "^3.2.0",
     "cssnano": "^4.1.10",
-    "eslint": "^5.16.0",
-    "eslint-loader": "^2.1.2",
-    "eslint-plugin-react": "^7.12.4",
-    "file-loader": "^4.0.0",
+    "eslint": "^6.6.0",
+    "eslint-loader": "^3.0.2",
+    "eslint-plugin-react": "^7.16.0",
+    "file-loader": "^4.2.0",
     "html-webpack-plugin": "^3.2.0",
-    "husky": "^2.4.1",
-    "less": "^3.9.0",
+    "husky": "^3.0.9",
+    "less": "^3.10.3",
     "less-loader": "^5.0.0",
-    "lint-staged": "^8.2.1",
-    "mini-css-extract-plugin": "^0.7.0",
-    "nodemon": "^1.19.1",
+    "lint-staged": "^9.4.3",
+    "mini-css-extract-plugin": "^0.8.0",
+    "nodemon": "^1.19.4",
     "postcss-loader": "^3.0.0",
-    "postcss-preset-env": "^6.6.0",
-    "prettier": "1.18.2",
-    "react-hot-loader": "^4.11.0",
+    "postcss-preset-env": "^6.7.0",
+    "prettier": "1.19.1",
+    "react-hot-loader": "^4.12.18",
     "redux-logger": "^3.0.6",
-    "rimraf": "^2.6.3",
-    "style-loader": "^0.23.1",
-    "terser-webpack-plugin": "^1.2.3",
-    "url-loader": "^2.0.0",
-    "webpack": "^4.34.0",
-    "webpack-bundle-analyzer": "^3.1.0",
-    "webpack-cli": "^3.3.4",
-    "webpack-dev-server": "^3.7.1",
-    "webpack-merge": "^4.2.1",
+    "rimraf": "^3.0.0",
+    "style-loader": "^1.0.0",
+    "terser-webpack-plugin": "^2.2.1",
+    "url-loader": "^2.2.0",
+    "webpack": "^4.41.2",
+    "webpack-bundle-analyzer": "^3.6.0",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0",
+    "webpack-merge": "^4.2.2",
     "zip-webpack-plugin": "^3.0.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.4.3",
-    "antd": "^3.20.0",
+    "@babel/runtime": "^7.7.2",
+    "antd": "^3.25.2",
     "axios": "^0.19.0",
-    "core-js": "^3.1.3",
-    "echarts": "^4.2.1",
+    "core-js": "^3.4.1",
+    "echarts": "^4.4.0",
     "fastclick": "^1.0.6",
-    "rc-form": "^2.4.5",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-intl-universal": "^2.1.1",
-    "react-pdf-js": "^5.0.5",
-    "react-redux": "^7.1.0",
-    "react-router-dom": "^5.0.1",
-    "react-signature-canvas": "^1.0.1",
-    "redux": "^4.0.1",
+    "rc-form": "^2.4.10",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-intl-universal": "^2.2.1",
+    "react-pdf-js": "^5.1.0",
+    "react-redux": "^7.1.3",
+    "react-router-dom": "^5.1.2",
+    "react-signature-canvas": "^1.0.3",
+    "redux": "^4.0.4",
     "redux-thunk": "^2.3.0"
   }
 }
