@@ -1,7 +1,7 @@
 /**
  * login
  */
-import React from "react";
+import React, { useState, useEffect } from "react";
 import { Redirect } from "react-router-dom";
 import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
@@ -11,107 +11,73 @@ import { Application } from "Utils/storage";
 import { Form, Icon, Input, Button } from "antd";
 import "./style";
 
-class Login extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      redirectToReferrer: props.token || false
-    };
-  }
+function Login(props) {
+  const [redirectToReferrer, setRedirectToReferrer] = useState(false);
 
-  componentDidMount() {}
+  useEffect(() => {
+    setRedirectToReferrer(props.token);
+  }, [props.token]);
 
-  /**
-   * login
-   * 登陆成功后需要获取菜单
-   */
-  fnLogin = async () => {
-    const data = {
-      account: "test",
-      password: "password"
-    };
-    await this.props.loginAction(data, () =>
-      this.setState({ redirectToReferrer: true })
-    );
-
-    // 加载资源中
-    await this.props.menuAction();
-  };
-
-  fnBack = () => {
-    Storage.clear();
-    goBack();
-  };
-
-  fnHandleSubmit = e => {
+  function fnLogin(e) {
     e.preventDefault();
-    this.props.form.validateFields(async (err, values) => {
+    props.form.validateFields(async (err, values) => {
       if (!err) {
         console.log("Received values of form: ", values);
 
-        await this.props.loginAction(values, () =>
-          this.setState({ redirectToReferrer: true })
-        );
+        await props.loginAction(values, () => setRedirectToReferrer(true));
 
-        // 加载资源中
-        await this.props.menuAction();
+        // load menu
+        await props.menuAction();
       }
     });
-  };
-
-  render() {
-    const { redirectToReferrer } = this.state;
-    if (redirectToReferrer) {
-      const { from } = this.props.location.state || {
-        from: { pathname: "/" } // default site
-      };
-      return <Redirect to={from} />;
-    }
-
-    const { getFieldDecorator } = this.props.form;
-    return (
-      <div className="login">
-        <Form onSubmit={this.fnHandleSubmit} className="login-form">
-          <div className="title">Admin</div>
-          <Form.Item>
-            {getFieldDecorator("username", {
-              rules: [{ required: true, message: "please input usename!" }],
-              initialValue: Application.get("username") || ""
-            })(
-              <Input
-                prefix={
-                  <Icon type="user" style={{ color: "rgba(0,0,0,.25)" }} />
-                }
-                placeholder="username"
-              />
-            )}
-          </Form.Item>
-          <Form.Item>
-            {getFieldDecorator("password", {
-              rules: [{ required: true, message: "please input password!" }]
-            })(
-              <Input
-                prefix={
-                  <Icon type="lock" style={{ color: "rgba(0,0,0,.25)" }} />
-                }
-                type="password"
-                placeholder="password"
-              />
-            )}
-          </Form.Item>
-          <Form.Item>
-            <Button
-              type="primary"
-              htmlType="submit"
-              className="login-form-button"
-            >
-              Login in
-            </Button>
-          </Form.Item>
-        </Form>
-      </div>
-    );
   }
+
+  if (redirectToReferrer) {
+    const { from } = props.location.state || {
+      from: { pathname: "/" } // default site
+    };
+    return <Redirect to={from} />;
+  }
+
+  const { getFieldDecorator } = props.form;
+  return (
+    <div className="login">
+      <Form onSubmit={fnLogin} className="login-form">
+        <div className="title">Admin</div>
+        <Form.Item>
+          {getFieldDecorator("username", {
+            rules: [{ required: true, message: "please input usename!" }],
+            initialValue: Application.get("username") || ""
+          })(
+            <Input
+              prefix={<Icon type="user" style={{ color: "rgba(0,0,0,.25)" }} />}
+              placeholder="username"
+            />
+          )}
+        </Form.Item>
+        <Form.Item>
+          {getFieldDecorator("password", {
+            rules: [{ required: true, message: "please input password!" }]
+          })(
+            <Input
+              prefix={<Icon type="lock" style={{ color: "rgba(0,0,0,.25)" }} />}
+              type="password"
+              placeholder="password"
+            />
+          )}
+        </Form.Item>
+        <Form.Item>
+          <Button
+            type="primary"
+            htmlType="submit"
+            className="login-form-button"
+          >
+            Login in
+          </Button>
+        </Form.Item>
+      </Form>
+    </div>
+  );
 }
 
 const mapStateToProps = state => ({
@@ -126,4 +92,4 @@ const mapDispatchToProps = dispatch => ({
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(Form.create({ name: "normal_login" })(Login));
+)(Form.create({ name: "login" })(Login));
